Change login method

+++ src/main/java/src/ICustomer.java
@@ -2,8 +2,8 @@ package src;
 
 public interface ICustomer extends IUser {
 
-    public void withdrawCash(int amount);
-    public void depositCash(int amount);
+    public void withdrawCash(double amount);
+    public void depositCash(double amount);
     public void displayBalance();
     public void exit();
